+++ backend/src/Web/ClientApp/src/app/app.component.html
@@ -5,7 +5,7 @@
   </p-progressBar>
   <app-nav-menu></app-nav-menu>
   <div class="container">
-    <app-page-alert></app-page-alert>
+    <p-toast></p-toast>
     <router-outlet></router-outlet>
   </div>
 </body>
