--- backend/src/Web/ClientApp/src/app/page-alert/page-alert.component.html
@@ -1,5 +1,0 @@
-<p *ngFor="let alert of alerts">
-  <ngb-alert [dismissible]="false" [type]="alert.type">
-    {{ alert.message }}
-  </ngb-alert>
-</p>
